requirements.txt
playwright==1.44.0
supabase==2.5.3
requests==2.32.3
python-dotenv==1.0.1
